Develop

+++ bli.py
@@ -20,6 +20,7 @@ def print_usage():
 	print "Options:"
 	print "  {0:20s} print downstream health".format("status")
 	print "  {0:20s} print downstream config".format("config")
+	print "  {0:20s} print downstream info (version, et al)".format("info")
 	print "  {0:20s} clear crash logs".format("clear_logs")
 	print "  {0:20s} check policy".format("check_policy")
 	print "  {0:20s} give this help list".format("-?, --help")
@@ -32,6 +33,7 @@ def menu():
 	print "|{0:28s}|".format(" (2) Print status")
 	print "|{0:28s}|".format(" (3) Clear crash logs")
 	print "|{0:28s}|".format(" (4) Check policy")
+	print "|{0:28s}|".format(" (8) Print information")
 	print "|{0:28s}|".format(" (9) Print configuration")
 	print "|{0:28s}|".format(" (0) Quit")
 	print "-"*30
@@ -48,9 +50,10 @@ def populate_sensors(sensor, sensors):
 
 		if line[0][0] != "#":
 			line[-1] = line[-1].strip()
-			sensors[line[0]] = {}
-			sensors[line[0]]['crash_logs']   = 0
+			sensors[line[0]] 		= {}
+			sensors[line[0]]['crash_logs']  = 0
 			sensors[line[0]]['policy_file'] = {}
+			sensors[line[0]]['version']	= {}
 			sensors[line[0]]['hostname']    = line[1]
 			sensors[line[0]]['ssh_user']    = line[2] if line[2] else ssh_user
 			sensors[line[0]]['prefix']      = line[3] if line[3] else prefix
@@ -106,6 +109,10 @@ def get_status(sensors):
 				line[1] = line[1].split(os.path.join(sensors[sensor]['prefix'], "share", "bro", "site/"))
 				sensors[sensor]['policy_file'][line[1][-1]] = line[0]
 
+			(stdin, stdout, stderr) = ssh.exec_command(os.path.join(sensors[sensor]['prefix'], "bin", "bro") + " --version 2>&1; " + os.path.join(sensors[sensor]['prefix'], "bin", "broctl") + " help |grep -i version 2>&1")
+			sensors[sensor]['version']['bro'] = stdout.readline().strip().split(' ')[-1]
+			sensors[sensor]['version']['broctl'] = stdout.readline().strip().split(' ')[-1]
+
 			if not fnf_prefix and not fnf_spool:
 				if running == lines:
 					sensors[sensor]['status'] = "OK (" + str(warnings) + " warnings, " + str(sensors[sensor]['crash_logs']) + " crash logs)"
@@ -133,28 +140,38 @@ def print_status(sensors):
 
 def clear_logs(sensors):
 	cleared = 0
+	purge = False
 
-	for sensor in sensors:
+	for sensor in sorted(sensors):
+		decision = None
 		if sensors[sensor]['crash_logs'] > 0:
-			ssh = paramiko.SSHClient()
-			ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+			if purge != True:
+				decision = raw_input("\nClear logs from "+sensor+"? ([y]es] / [n]o / [a]ll) ")
+				if decision == "a":
+					purge = True
 
-			try:
-				ssh.connect(
-					sensor,
-					username = sensors[sensor]['ssh_user'],
-					key_filename = os.path.expanduser(os.path.join("~", ".ssh", "id_rsa.pub")),
-					timeout = 10
-				)
-				(stdin, stdout, stderr) = ssh.exec_command("rm -rf " + sensors[sensor]['spooltmp'] + "/*crash")
-				ssh.close()
-				cleared += 1
-				print "\nLogs cleared from", sensor + "..."
-			except Exception as e:
-				sensors[sensor]['status'] = "Error (" + str(e) + ")"
+			if purge == True or decision == "y":
+				ssh = paramiko.SSHClient()
+				ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+
+				try:
+					ssh.connect(
+						sensor,
+						username = sensors[sensor]['ssh_user'],
+						key_filename = os.path.expanduser(os.path.join("~", 	".ssh", "id_rsa.pub")),
+						timeout = 10
+						)
+					(stdin, stdout, stderr) = ssh.exec_command("rm -rf " + sensors[sensor]['spooltmp'] + "/*crash")
+					ssh.close()
+					cleared += 1
+					print "\nLogs cleared from", sensor + "..."
+				except Exception as e:
+					sensors[sensor]['status'] = "Error (" + str(e) + ")"
+			elif decision == "n":
+				print "\nSkipping", sensor+"..."
 
 	if cleared == 0:
-		print "\nNo log(s) cleared..."
+		print "\nNo logs cleared."
 
 def check_policy(sensors):
 	stdout = subprocess.Popen(
@@ -202,16 +219,28 @@ def check_policy(sensors):
 						for policy_file in policy[sensors[sensor]['policy_type']]:
 							if policy_file in sensors[sensor]['policy_file']:
 								if policy[sensors[sensor]['policy_type']][policy_file] != sensors[sensor]['policy_file'][policy_file]:
-									print "{0:15s} : {1:20s} : {2:7s}: {3:8s} : {4} ".format("", "", "", "modified", policy_file)
+									print "{0:15s} : {1:20s} : {2:6s} : {3:8s} : {4} ".format("", "", "", "modified", policy_file)
 							else:
 								if policy_file != "error":
-									print "{0:15s} : {1:20s} : {2:7s} : {3:8s} : {4} ".format("", "", "", "missing", policy_file)
+									print "{0:15s} : {1:20s} : {2:6s} : {3:8s} : {4} ".format("", "", "", "missing", policy_file)
 					else:
 						print "{0:15s} : {1:20s} : {2:6s} : {3:8s} : {4} ".format("", "", "", "error", "policy '"+sensors[sensor]['policy_type']+"' not defined in bli configuration")
 				else:
 					print "{0:15s} : {1:20s} : {2:6s} : {3:8s} : {4} ".format("", "", "", "error", "deployment information for policy '"+sensors[sensor]['policy_type']+"' does not exist")
+			else:
+				print "{0:15s} : {1:20s} : {2:6s} : {3:8s} : {4}".format(sensor, sensors[sensor]['hostname'], "--", "--", "--")
 	else:
 		print "\nPolicy validation unavailable due to lack of deployment data"
+
+def print_info(sensors):
+	print "\n{0:15s} : {1:20s} : {2:11s} : {3:10s}".format("IP Address", "Hostname", "Bro Version", "Broctl Version")
+	print "-"*120
+
+	for sensor in sorted(sensors):
+		if "Error" not in sensors[sensor]['status']:
+			print "{0:15s} : {1:20s} : {2:11s} : {3:10s}".format(sensor, sensors[sensor]['hostname'], sensors[sensor]['version']['bro'], sensors[sensor]['version']['broctl'])
+		else:
+			print "{0:15s} : {1:20s} : {2:11s} : {3:10s}".format(sensor, sensors[sensor]['hostname'], "--", "--")
 
 def main():
 	loaded = decision = None
@@ -223,7 +252,7 @@ def main():
 			menu()
 
 			try:
-				decision = input("\naction> ")
+				decision = int(raw_input("\naction> "))
 			except Exception as e:
 				decision = None
 
@@ -246,6 +275,9 @@ def main():
 					elif decision == 4:
 						check_policy(sensors)
 						raw_input("\n<Press Enter to continue>")
+					elif decision == 8:
+						print_info(sensors)
+						raw_input("\n<Press Enter to continue>")
 				else:
 					print "\nStatus not yet loaded (get status)"
 
@@ -255,6 +287,9 @@ def main():
 			print_status(sensors)
 		elif sys.argv[1] == "config":
 			print_config(sensors)
+		elif sys.argv[1] == "info":
+			get_status(sensors)
+			print_info(sensors)
 		elif sys.argv[1] == "clear_logs":
 			get_status(sensors)
 			clear_logs(sensors)
